bus: mhi: Fix FW loading time delay
The QRTR TX CTRL packets sent from the device to Q6
timesout waiting for the ackowledgemnt packet from
Q6 in the MHI layer.
The timeout of 5 secs for each CTRL packet sent results
in the CNSS FW READY timeout and it asserts.

So, as an hack the CNSS FW READY timeout was increased
by setting the enviroinment variables which opened up the
delay issue for the wifi radio to come up.

This patch fixes the data corruption becuase of the endianness
and makes sure the Q6 generates the interrupt.

Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>
Change-Id: I8cd5341b32236eb3bc89042f027990dbc8d5f17c
(cherry picked from commit b81c7051cd3cda30343cfbab83438be3b38b4aa9)
Signed-off-by: Praveenkumar I <quic_ipkumar@quicinc.com>