mhi: netdev:  Avoid duplicate syncup before queing the buffers
Avoid additional invalidate on the mhi buffers before queing them
to the hardware rings. This is with the assumption that , network
stack does not make the buffers dirty, so they can be directly queued.
This saves about 10% cpu utilization during the overall data testing
at about 6Gbps. More the data rate, more would be the savings.

NOTE: Incase, if buffers were made dirty for any reason by nw stack,
this invalidate cannot be avoided and this commit has to be reverted
back.

Change-Id: Ib8c3cbe413a936d2e0673d0126a20623bf4b34a5
Signed-off-by: Sricharan R <quic_srichara@quicinc.com>