arm64: dts: qcom: devsoc: coresight: specify the ATID for the ETM source
Specify the ATID for all the ETM source. ATID for the ETM source
starts from 1. Upon selection of the source, corresponding ATID will
be configured in the ATID filtering block, so that trace data can be
routed to ETR.

Also, specify the CSR ETR ATID register offset to configure the ATID
into it.

Change-Id: I5e8657089c4ed27ea71461374a581e8d0f6934d9
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>