mhi: Changed IRQ_HANDLED to IRQ_NONE in mhi_irq_handler
This patch changes the IRQ_HANDLED return value to IRQ_NONE,
if there is no pending event in the event ring to be handled.

In case of shared interrupt, kernel will invokes all handlers
registered for the particular interrupt with unique dev_id(in our
case it is mhi_event), and its shared handlers responsibility to 
identify its own interrupt and exits quickly if not.

In our case, we are exiting the mhi_irq_handler if there is no pending
events in that particular event ring. So the return ideally should be
IRQ_NONE which denotes interrupt was not from this device or was not handled.

Change-Id: Ia2d8a125c8663e75b4ee3f0dc8aec5fdd1283b41
Signed-off-by: Karthick Shanmugham <quic_kartshan@quicinc.com>