arm: change textofs for devsoc 256MB profile
For 256M profiles, based on the LOAD_ADDR, we have defined the
TEXT_OFFSET to calculate the correct DDR base.

For devsoc, LOAD_ADDR is changed to DDR Base + TEXT_OFFSET, since
there is no reserved memory at the start of the DDR. Make necessary
changes to rely on the default TEXT_OFFSET for the devsoc platform.

Change-Id: Ie843e701071dcb09af2e89f6db3298b68435dffe
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>