arm64: dts: devsoc: add the PCIe PHY support for one dual lane / two single lane
PCIE1 is a dual lane controller, PCIE2 is a single lane controller both
shares the same PHY. 1st lane of PCIE1 is connected to 1st lane in the
PHY and there is a mux to select between the 2nd lane in PCIE1 to make
it as a dual lane controller or we enable the PCIE2 so that both PCIE1
and PCIE2 works simultaneously in a single lane mode. Pictorial
representation of this configuration can be found at the commit
f73af5c4d60a ("arm64: dts: qcom: devsoc: add the PCIe support")

Each PHY configuration has its own clocks and resets. So added a two single
lane phy nodes to include the correspoding clocks and resets. We can
enable both these PHYs when we need PCIE1 and PCIE2 to be enabled.

Also added a separate dual lane phy which can be enabled when we need
the PCIE1 to work in dual lane mode.

Change-Id: I183bdc586d41888f43cf0371f7e4825a524f9821
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>