arm64: dts: qcom: devsoc: add the PHY nodes for the PCIe controllers
By default, PCIE1 will be in dual lane mode, so associate the dual lane
phy by default. If we need to change it to single lane, we can modify
the 'phys' property in corresponding RDP DTS files.

While at it, update the PCIe phandle names to align with the IPCAT
naming convention for better readability.

Change-Id: Ieeb4d08ac95636efbb7c9fbb72fdffcb3b4eba9b
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>