pcie: qcom: add support to read axi_m clk freq from DT
AXI_M_CLK frequency will be different for single lane and dual lane.
Add support to read the frequency from DT and set the clock rate
accordingly.

In IPQ5332, by default PCIE1 will be in dual lane mode. So specify the
clock frequency which is needed for dual lane mode in SoC DTS. If the
PCIE1 is changed to single lane mode, we can update the clock frequency
in that board specific DTS files.

While at it, introduce a separate compatible for IPQ5332. ops is similar
to IPQ9574, so re-use the same.

Change-Id: I9687a4e0fbf7541131c6d6fdebd2a44bbd5d2970
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>