cpufreq: qti: Enable cpufreq for ipq53xx
IPQ53xx have different OPPs available for the CPU based on
SoC variant. This can be determined through use of an eFuse
register present in the silicon.

Added support for ipq53xx on nvmem driver which helps to
determine OPPs at runtime based on the eFuse register which
has the CPU frequency limits. opp-supported-hw dt binding
can be used to indicate the availabe OPPs for each limit.

nvmem driver also creates the "cpufreq-dt" platform_device after
passing the version matching data to the OPP framework so that the
cpufreq-dt handles the actual cpufreq implementation.

Currently, IPQ53xx SKUs supports either 1.1GHz or 1.5GHz based on the
fuse, which means no dynamic frequency scaling as well. If no fuse is
blown, 1.5GHz will be selected for now.

Change-Id: I8e5ca7637ccc0e56713a368cd005f5125099b9b0
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>