crypto: seccrypt: Qualcomm seccrypt driver
This driver is mainly used for SoCs which is not having
hardware crypto engine.This driver will fallback to arm-isa
cryptography or trustzone for encryption/decryption.

-fallback trustzone:
 specify the property "seccrypt,fallback_tz" in device tree.

-skcipher algorithms
 implementation of AES crypto API, the crypto register alg function.

-cipher mode support
 ECB, CBC, CTR.

Since if there is no hardware crypto engine present in SoCs , so
the encryption/decryption will be done by trustzone via SMC call
or arm-isa cryptography instruction.

Change-Id: Ife0e6072500fae8f4733b134d346829172985aef
Signed-off-by: Md Sadre Alam <quic_mdalam@quicinc.com>