firmware: Modified fuseipq scm call return value
The return value of fuseipq scm call is getting overwritten with
the response status from secure world. Because of this all other
error cases are masked and same error will be printed always.
And positive return value of sysfs is triggering store call again
assuming data was partially written.

In this patch, removed the fuseipq scm call overwrite and returned as it
is. fuse_status will hold the return status code coming from secure world.

Change-Id: Ia0a710ca34316e68f13c685b0a52cd093ceef42c
Signed-off-by: Karthick Shanmugham <quic_kartshan@quicinc.com>