mhi: test: use MHITEST_VERB in mhitest_dump_info
This patch changes the prints in mhitest_dump_info function
to MHITEST_VERB. This function is called in the panic path
and when dump size is larger, more prints are coming via
MHITEST_EMERG which is causing watch dog bite. Hence this
patch changes the log level to prevent the watchdog bite and
dump info logs can be enabled when debug_lvl is set to 0.

Change-Id: I34d7d7976b9f8cb1e9856f76ac92f23b15c70d8a
Signed-off-by: Saahil Tomar <quic_saahtoma@quicinc.com>