arm64: dts: Optimize memory reserved for IPQ53xx Q6.
Reduce 8MB from the memory reserved for IPQ53xx Q6.
Retain the current MLO start address as it's hard coded in
FW.
Also, in QCN9160 RDP, exclude MLO from the overall Q6 reserved
memory. It's not exclusive to IPQ53xx, it is in fact a shared
memory region between IPQ53xx and QCN9224.

Change-Id: I113eefc24c83c70ce987f0e3c88dcfb86e6d4b58
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>