dts: ipq5018: Disable system_noc rate adapter due to regression
Disable system_noc rate adapter in the pcie driver. The initial
implementation aimed to configure the rate adapter register based
on a calculated value to optimize bandwidth utilization. However,
it has been discovered that this change leads to a regression and
a subsequent 15% throughput drop on SDX peripheral, particularly
impacting mp03.6.

Change-Id: Iaa9f75a2e9e8025502b74c55a1950f57513e3884
Signed-off-by: Amandeep Singh <quic_amansing@quicinc.com>